Fix up arguments to logging methods.
We have a special logging formatter that is used to show info when a
django HttpRequest is part of the log record. Once upon a time, the way
of getting extra information into a log record was just to pass it as
additional `kwargs` to the logging methods, but the modern, documented
way of doing it is to pass a single `extra` dict. Some of our code uses
`extra`, but most of it is just passing in `request=` as a kwarg. I've
fixed this up, plus added the request in logging methods where we had it
available but weren't plumbing it.

In addition, there were a couple instances where we were passing
`exc_info=1` instead of `exc_info=True`. I've fixed these up to make
type checkers happier.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12704/