Allow the use of typescript (.ts) files.
This change adds the babel preset for typescript and a babel compiler
that matches it. With this, `.ts` files can be added to JavaScript
bundles and will work correctly.

I've stuck a tsconfig.json file in the root. This isn't critical because
we don't use tsc to compile .ts to .js (that job is handled by babel),
but this allows us to use things like tsserver for the dev environment
without getting errors about non-ES3 constructs.

Testing Done:
Used some `.ts` files in the Review Board codebase.

Reviewed at https://reviews.reviewboard.org/r/12727/