Fix ListEditView JS tests.
The JS tests for the ListEditView had a couple bugs. First was that we
had a mix of string characters, with both backtick strings and
single-quote strings that were concatenated. This meant that the
generated `<input>` element had some garbage characters in it with extra
quotes and + signs.

Second was that the default row HTML we were creating was only being passed
to the template, and not also to the view.

Testing Done:
Ran js-tests and saw all djblets tests pass.

Reviewed at https://reviews.reviewboard.org/r/12815/