Update the extlinks configuration for our release notes for newer Sphinx.
Modern versions of Sphinx require some updates to `extlinks`
configuration, requiring a `%s` in the extlink title, so it can fill in
a value at an explicit spot.

This change adds this to all of our titles, mirroring the values we have
in Review Board.

Testing Done:
Successfully built the docs on modern Sphinx.

Reviewed at https://reviews.reviewboard.org/r/12842/