Fix an infinite loop if rollup config can't be found.
The code to try to find the rollup configuration file iterates up
the directory tree until either the config is found or we hit the root.
Unfortunately, this was comparing a `Path` object (for the current path)
against `str` (for the root), which always evaluates to `False`. The
result was that if there's no rollup config, we'd end up infinitely
looping comparing `Path('/')` to `/`. This change fixes that.

Testing Done:
Tried to load an extension which defined an `index.ts` file in its
bundles but did not have a `rollup.config.js` file. Saw that we no
longer hit the infinite loop.

Reviewed at https://reviews.reviewboard.org/r/12956/