Add type hints, missing docs, enums, and cleanup to djblets.features.
This adds missing type hints to `djblets.features`, and in the process
updates some of the code to add missing docstrings and to fix types in
some docstrings.

`FeatureLevel` is now an enum, which should be backwards-compatible
with the original `int` in how this is used. That gives better type
checking than using an `int`.

`FeatureRegistry` has been updated to be a typed registry.

Testing Done:
Unit tests pass.

Tested with some code defining a new feature.

Reviewed at https://reviews.reviewboard.org/r/12991/