Add pytest mode to djblets.testing.testrunners.TestRunner.
We've mostly phased out nose from our own usage, but it's still being
used for the extension test runner. The behavior is different enough
that we need a transition period to move people over.

In most cases, once extensions are configured in a way to make use of
pytest, they'll be able to just run `pytest` instead of `rbext test`,
but we also want latter to continue to exist.

This change adds a mode to the TestRunner to use pytest instead of nose.
That can be opted into for now, and in a future version we'll switch it
over and remove the nose mode.

Testing Done:
Ran rbintegrations tests via pytest and our test runner.

Reviewed at https://reviews.reviewboard.org/r/13075/