Update Djblets to use Housekeeping for deprecation management.
This adds a dependency on housekeeping, our new deprecation management
library, and switches all of our base classes over to it.

The original base class and deprecation utility functions are still
there, but wrap the versions in housekeeping. They themselves are marked
as deprecated and will be removed in Djblets 6.

Housekeeping 1.1 is needed in order to allow us to retain
`BaseRemovedInProductVersionWarning` as a base class (until removed in
Djblets 6) while also marking it deprecated and avoiding a warning in
the `PendingRemovalInDjbletsWarning` subclass.

Testing Done:
Unit tests pass in Djblets and Review Board.

Reviewed at https://reviews.reviewboard.org/r/13124/