Update Djblets to use Housekeeping for deprecation management.
This adds a dependency on housekeeping, our new deprecation management
library, and switches all of our base classes over to it.

The original base class and deprecation utility functions are still
there, but wrap the versions in housekeeping. They themselves are marked
as deprecated and will be removed in Djblets 6.

The `BaseRemovedInProductVersionWarning` is deprecated, but remains for
now. We can't mark it as moved or deprecated using housekeeping, given
circular references (which Housekeeping 1.1 was supposed to fix, but
ultimately couldn't).

Testing Done:
Unit tests pass in Djblets and Review Board.

Reviewed at https://reviews.reviewboard.org/r/13124/