Port config forms JS codebase to TypeScript/spina.
This change ports the JavaScript side of the config forms app to use
TypeScript and spina.

For the most part this is a very straightforward port. The only annoying
thing was I discovered that the `ListItem` implementation allowed
passing in an `actions` parameter which was treated as an option arg,
but was actually being passed in as part of the attributes (despite not
being an attribute on the model). This was only used inside of our unit
tests (which have themselves been updated to use `setActions`), but I've
added some compatibility code to emit a deprecation warning just in the
very unlikely case that someone was using this.

Testing Done:
Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/13140/