Add padding to size calculation for modal box contents.
The review dialog in Review Board has had a long-standing bug where the
buttons could overlap with the dialog contents. This was due to the size
calculation including margin but not padding for the modal box contents.
This change adds that.

Testing Done:
Verified that the dialog contents no longer overlapped the buttons.

Reviewed at https://reviews.reviewboard.org/r/13172/