Improve typing support for the UNSET symbol.
This reworks the internals of the UNSET symbol (which was introduced in
Djblets 3.3) and makes it easier to integrate it with typing logic and
type checkers.

`UnsetSymbol` is now an `Enum`, allowing `UNSET` to be defined as a
`Literal`. This works along with a new `Unsettable[_T]` type alias to
mark whether a parameter or variable allows `UNSET` to be set.

By making this a `Literal`, type checkers can type narrow when a code
compares a value to `UNSET`, avoiding any type errors that still wanted
to consider `UnsetSymbol` a possible value.

Testing Done:
Made use of this in some in-progress code. Verified that mypy and pyright
successfully narrowed the types when comparing against `UNSET`.

Reviewed at https://reviews.reviewboard.org/r/13191/