Add a health check view.
This change adds a view implementation that can be used to set up a
health check. It simply checks that the database and cache server are
available, and returns 200 if so and 500 if not (while logging the
error).

Testing Done:
Manually set things up to have either the database or cache server fail
and saw that this view returned the correct results.

Reviewed at https://reviews.reviewboard.org/r/13252/