Added type hints and modernize docs for djblets.webapi.errors.
This updates `djblets.webapi.errors` to include type hints and modern
documentation. This class is ancient, and was not up to our current
standards.

This is in preparation for upcoming work on error support.

There is one small feature addition in that `with_override()` and
`with_message()` will now do the right thing if called on a `WebAPIError`
subclass. While not a pattern we currently explicitly support, this is
more correct and helps future-proof the class.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13258/