Add typing and docs for WebAPIResource configuration attributes.
This updates `WebAPIResource` to add typing and documentation for the
attributes intended to be filled in when subclassing. This can help with
ensuring that subclasses properly configure new API resources.

There is the usual issue where some attributes (`name`, `name_plural`,
and some others) are defined as properties by default, and overriding
them causes issues with the type checker. This is a wider problem in how
we manage configuring subclasses, and won't be addressed as part of this
work.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13276/