Add and modernize documentation throughout WebAPIResource.
`WebAPIResource` is a very old and complex part of our codebase, and
lacked any form of modern documentation. This made it more difficult to
iterate on, and with some upcoming work planned for this, it was time to
address this problem.

This change adds documentation to all the methods in the class, and sets
the stage for those upcoming changes.

Testing Done:
Unit tests pass.

Built the docs. Verified there were no new errors in this module.

Reviewed at https://reviews.reviewboard.org/r/13277/