Update dependency minimums for Djblets.
We have a handful of dependencies for which we had old minimums,
including a few which were triggering security alerts. While none of
these actually impacted the security of the product, we still want to
start with a good baseline.

This change updates a few Python and JavaScript dependency versions to
fix these issues.

Testing Done:
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13330/