Add typing and fix issues throughout the datagrids code.
The datagrids code was old and crufty, and made some bad typing
assumptions all throughout. A few of these have bit us in the past, when
search indexing bots threw garbage at our datagrids and caused HTTP 500
errors. The mess of types have made it hard to reliably work on this
code, and since it needs some upcoming work, it was time to fix that up.

This change adds typing and documentation all throughout the datagrids
code. In the process, bad typing assumptions and scope errors were
fixed.

There's a lot of old, bad design patterns in this code, but this doesn't
try to change any of those. It's just building enough of a foundation to
start iterating on.

Testing Done:
Djblets and Review Board unit tests pass.

Tested all of the datagrids used in Review Board.

Reviewed at https://reviews.reviewboard.org/r/13355/